+++ packages/plugin-reanimated/tsdown.config.ts
@@ -1,3 +1,7 @@
 import baseConfig from '@repo/configs/tsdown.config';
+import { defineConfig } from 'tsdown';
 
-export default baseConfig;
+export default defineConfig({
+  ...baseConfig,
+  external: ['react-native-reanimated-original'],
+});